+++ cmd/cycloid/middleware/middleware.go
@@ -39,7 +39,7 @@ type Middleware interface {
 	Login(email, password string) (*models.UserSession, error)
 
 	// LoginOrg is the used to log the user into a Cycloid organization
-	LoginOrg(org, email, password string) (*models.UserSession, error)
+	LoginOrg(org, child, email, password string) (*models.UserSession, error)
 
 	DeleteMember(org string, name string) error
 	GetMember(org string, name string) (*models.MemberOrg, error)